+++ src/main/kotlin/no/nav/bidrag/behandling/transformers/grunnlag/Grunnlagsmapping.kt
@@ -173,7 +173,7 @@ fun Behandling.opprettAldersjusteringDetaljerGrunnlag(
     aldersjustert: Boolean = true,
     søknadsbarn: Rolle,
     begrunnelser: List<String>? = null,
-    vedtaksidBeregning: Long? = null,
+    vedtaksidBeregning: Int? = null,
 ) = GrunnlagDto(
     referanse = "${no.nav.bidrag.domene.enums.grunnlag.Grunnlagstype.ALDERSJUSTERING_DETALJER}_${tilStønadsid(
         søknadsbarn,
